chore: Refactor some smaller test parts
