+++ phpstan.neon.dist
@@ -127,7 +127,6 @@ parameters:
         - # tests are allowed to add dummy classes in the same file
             message: '#Multiple class\/interface\/trait is not allowed in single file#'
             paths:
-                - src/**/Test/**
                 - tests/**
 
         # implementation of symfony flashbag interface which does not define iterable types
