+++ src/Core/Framework/Test/DataAbstractionLayer/Field/TestDefinition/SubCascadeDefinition.php
@@ -1,15 +1,18 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Shopware\Core\Framework\Test\DataAbstractionLayer\Field\TestDefinition;
 
 use Shopware\Core\Framework\DataAbstractionLayer\EntityDefinition;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\FkField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\ApiAware;
-use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\CascadeDelete;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\PrimaryKey;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\Required;
-use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\RestrictDelete;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\IdField;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\IntField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\OneToOneAssociationField;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\ReferenceVersionField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\StringField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\VersionField;
 use Shopware\Core\Framework\DataAbstractionLayer\FieldCollection;
@@ -17,9 +20,9 @@ use Shopware\Core\Framework\DataAbstractionLayer\FieldCollection;
 /**
  * @internal
  */
-class RootDefinition extends EntityDefinition
+class SubCascadeDefinition extends EntityDefinition
 {
-    final public const ENTITY_NAME = 'root';
+    final public const ENTITY_NAME = 'root_sub_cascade';
 
     public function getEntityName(): string
     {
@@ -28,7 +31,7 @@ class RootDefinition extends EntityDefinition
 
     public function since(): ?string
     {
-        return '6.0.0.0';
+        return '6.3.3.0';
     }
 
     protected function defineFields(): FieldCollection
@@ -37,8 +40,10 @@ class RootDefinition extends EntityDefinition
             (new IdField('id', 'id'))->addFlags(new ApiAware(), new PrimaryKey(), new Required()),
             new VersionField(),
             new StringField('name', 'name'),
-            (new OneToOneAssociationField('sub', 'id', 'root_id', SubDefinition::class))->addFlags(new ApiAware(), new RestrictDelete()),
-            (new OneToOneAssociationField('subCascade', 'id', 'root_id', SubCascadeDefinition::class))->addFlags(new ApiAware(), new CascadeDelete()),
+            new IntField('stock', 'stock'),
+            new FkField('root_id', 'rootId', RootDefinition::class, 'id'),
+            (new ReferenceVersionField(RootDefinition::class))->addFlags(new ApiAware(), new Required()),
+            new OneToOneAssociationField('root', 'root_id', 'id', RootDefinition::class, false),
         ]);
     }
 }
