+++ src/Core/Framework/Test/DataAbstractionLayer/Field/TestDefinition/SubManyDefinition.php
@@ -1,15 +1,17 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Shopware\Core\Framework\Test\DataAbstractionLayer\Field\TestDefinition;
 
 use Shopware\Core\Framework\DataAbstractionLayer\EntityDefinition;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\FkField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\ApiAware;
-use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\CascadeDelete;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\PrimaryKey;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\Required;
-use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\RestrictDelete;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\IdField;
-use Shopware\Core\Framework\DataAbstractionLayer\Field\OneToOneAssociationField;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\ManyToOneAssociationField;
+use Shopware\Core\Framework\DataAbstractionLayer\Field\ReferenceVersionField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\StringField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\VersionField;
 use Shopware\Core\Framework\DataAbstractionLayer\FieldCollection;
@@ -17,9 +19,9 @@ use Shopware\Core\Framework\DataAbstractionLayer\FieldCollection;
 /**
  * @internal
  */
-class RootDefinition extends EntityDefinition
+class SubManyDefinition extends EntityDefinition
 {
-    final public const ENTITY_NAME = 'root';
+    final public const ENTITY_NAME = 'root_sub_many';
 
     public function getEntityName(): string
     {
@@ -28,7 +30,7 @@ class RootDefinition extends EntityDefinition
 
     public function since(): ?string
     {
-        return '6.0.0.0';
+        return '6.3.3.0';
     }
 
     protected function defineFields(): FieldCollection
@@ -37,8 +39,9 @@ class RootDefinition extends EntityDefinition
             (new IdField('id', 'id'))->addFlags(new ApiAware(), new PrimaryKey(), new Required()),
             new VersionField(),
             new StringField('name', 'name'),
-            (new OneToOneAssociationField('sub', 'id', 'root_id', SubDefinition::class))->addFlags(new ApiAware(), new RestrictDelete()),
-            (new OneToOneAssociationField('subCascade', 'id', 'root_id', SubCascadeDefinition::class))->addFlags(new ApiAware(), new CascadeDelete()),
+            (new FkField('root_sub_id', 'subId', SubDefinition::class, 'id'))->addFlags(new ApiAware(), new Required()),
+            (new ReferenceVersionField(SubDefinition::class))->addFlags(new ApiAware(), new Required()),
+            new ManyToOneAssociationField('sub', 'root_sub_id', SubDefinition::class, 'id', false),
         ]);
     }
 }
