+++ tests/migration/Core/MigrationForeignDefaultLanguageTest.php
@@ -14,8 +14,8 @@ use Shopware\Core\Framework\Migration\MigrationCollectionLoader;
 use Shopware\Core\Framework\Test\TestCaseBase\DatabaseTransactionBehaviour;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
 use Shopware\Core\Framework\Uuid\Uuid;
-use Shopware\Core\Migration\Traits\MigrationUntouchedDbTestTrait;
 use Shopware\Core\Migration\V6_3\Migration1536233560BasicData;
+use Shopware\Tests\Migration\MigrationUntouchedDbTestTrait;
 
 /**
  * @internal
